MathStudy

This is the mandatory index.md file that is the homepage of the documentation.

Here the documentation string of a function is expanded:

param_line — Function

param_line(p::Point, q::Point, n::Int64) → Ps::[Point]

Creates n points on a line defined by p and q, using the parametric equation of a line, then plots

We can link to any subsection name with Subsection name.

We can link to documentation strings with param_line.

Subsection name

Test.